refactor: move bulk upload data classes to separate files
